requirements.txt
ansible-core==2.13.7
astroid==2.12.14
boto3==1.26.45
botocore==1.29.45
cffi==1.15.1
cryptography==39.0.0
dill==0.3.6
isort==5.11.4
Jinja2==3.1.2
jmespath==1.0.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
mccabe==0.7.0
packaging==22.0
platformdirs==2.6.2
pycparser==2.21
pylint==2.15.9
python-dateutil==2.8.2
PyYAML==6.0
resolvelib==0.8.1
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
urllib3==1.26.13
wrapt==1.14.1
